+++ src/ByteDev.Common/Time/DateTimeOffsetProvider.cs
@@ -2,22 +2,10 @@
 
 namespace ByteDev.Common.Time
 {
-    public interface IDateTimeOffsetProvider
-    {
-        DateTimeOffset Now { get; }
-        DateTimeOffset UtcNow { get; }
-    }
-
     public class DateTimeOffsetProvider : IDateTimeOffsetProvider
     {
-        public DateTimeOffset Now
-        {
-            get { return DateTimeOffset.Now; }
-        }
+        public DateTimeOffset Now => DateTimeOffset.Now;
 
-        public DateTimeOffset UtcNow
-        {
-            get { return DateTimeOffset.UtcNow; }
-        }
+        public DateTimeOffset UtcNow => DateTimeOffset.UtcNow;
     }
 }