+++ src/ByteDev.Common/Time/IDateTimeOffsetProvider.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ByteDev.Common.Time
+{
+    public interface IDateTimeOffsetProvider
+    {
+        DateTimeOffset Now { get; }
+
+        DateTimeOffset UtcNow { get; }
+    }
+}